feat: add memberDeleteAdmin component/ add app_setting
